style fixes for cpp

+++ ModifiedFib/cpp/main.cpp
@@ -4,8 +4,8 @@ uint32_t sequence[25];
 
 int main()
 {
-    sequence[0] = static_cast<uint32_t>(0);
-    sequence[1] = static_cast<uint32_t>(1);
+    sequence[0] = 0;
+    sequence[1] = 1;
     sequence[2] = sequence[1];
     
     for(uint8_t i = 3; i < 25; i++)
